requirements.txt
flask
python-dotenv
requests
praw